修改价格为float类型，现在价格可以为小数了。 增加了api拦截器，提高了系统的安全系数。 修复了account页面的数据显示错误

+++ src/main/java/com/example/dao/ProductDAO.java
@@ -89,5 +89,5 @@ public interface ProductDAO {
      */
 
 
-    Integer getPriceById(Integer pid);
+    Float getPriceById(Integer pid);
 }
